Add files via upload
There dosent seem to be many arbitrage opportunities at the moment that net more than a few cents of profit. Currently, data is only pulled for the highest level of the orderbook for each coin due to api rate limits, if binance adds a grouping function to the orderbook it may be easier to identify profitable arbitrage opportunities. Future updates might include auto pulling api keys and visualizations using dash with historical data.